+++ src/init/index.ts
@@ -1,0 +1,3 @@
+export * from "./commands";
+export * from "./resources";
+export * from "./ui";
