+++ src/init/resources.ts
@@ -1,0 +1,118 @@
+import * as fs from "fs";
+import * as path from "path";
+import * as vscode from "vscode";
+
+import { Err, Ok, Result } from "ts-results";
+import Resources from "../config/resources";
+import { downloadFile, isJavaPresent } from "../utils";
+import {
+    TMC_JAR,
+    WORKSPACE_ROOT_FILE,
+    WORKSPACE_ROOT_FILE_TEXT,
+    WORKSPACE_SETTINGS,
+} from "../config/constants";
+
+/**
+ * Checks if Java is present and performs resource initialization on extension activation
+ * @param extensionContext Extension context
+ */
+export async function resourceInitialization(
+    extensionContext: vscode.ExtensionContext,
+): Promise<Result<Resources, Error>> {
+    if (!(await isJavaPresent())) {
+        return new Err(new Error("Java not found or improperly configured."));
+    }
+
+    const extensionVersion = vscode.extensions.getExtension("tmc-vscode-temporary.tmc-vscode")
+        ?.packageJSON.version;
+
+    const cssPath = extensionContext.asAbsolutePath("resources/styles");
+    const htmlPath = extensionContext.asAbsolutePath("resources/templates");
+    const mediaPath = extensionContext.asAbsolutePath("media");
+
+    const basePath = extensionContext.globalStoragePath;
+    const tmcDataPath = path.join(basePath, "tmcdata");
+    const tmcWorkspacePath = path.join(tmcDataPath, "TMC workspace");
+    const tmcWorkspaceFilePath = path.join(tmcWorkspacePath, "TMC Exercises.code-workspace");
+    const tmcExercisesFolderPath = path.join(tmcWorkspacePath, "Exercises");
+    const tmcClosedExercisesFolderPath = path.join(tmcDataPath, "closed-exercises");
+
+    const tmcLangsPath = path.join(tmcDataPath, "tmc-langs.jar");
+
+    if (!fs.existsSync(basePath)) {
+        fs.mkdirSync(basePath);
+        console.log("Created global storage directory at", basePath);
+    }
+
+    if (!fs.existsSync(tmcDataPath)) {
+        fs.mkdirSync(tmcDataPath);
+        console.log("Created tmc data directory at", tmcDataPath);
+    }
+
+    if (!fs.existsSync(tmcWorkspacePath)) {
+        fs.mkdirSync(tmcWorkspacePath);
+        console.log("Created tmc workspace directory at", tmcWorkspacePath);
+    }
+
+    if (!fs.existsSync(tmcWorkspaceFilePath)) {
+        fs.writeFileSync(tmcWorkspaceFilePath, JSON.stringify(WORKSPACE_SETTINGS));
+        console.log("Created tmc workspace file at", tmcWorkspaceFilePath);
+    }
+
+    if (!fs.existsSync(tmcExercisesFolderPath)) {
+        fs.mkdirSync(tmcExercisesFolderPath);
+        console.log("Created tmc exercise directory at", tmcExercisesFolderPath);
+    }
+
+    if (!fs.existsSync(path.join(tmcExercisesFolderPath, WORKSPACE_ROOT_FILE))) {
+        fs.writeFileSync(
+            path.join(tmcExercisesFolderPath, WORKSPACE_ROOT_FILE),
+            WORKSPACE_ROOT_FILE_TEXT,
+        );
+        console.log("Wrote tmc root file at", tmcExercisesFolderPath);
+    }
+
+    if (!fs.existsSync(tmcClosedExercisesFolderPath)) {
+        fs.mkdirSync(tmcClosedExercisesFolderPath);
+        console.log("Created tmc closed exercise directory at", tmcClosedExercisesFolderPath);
+    }
+
+    if (!fs.existsSync(tmcLangsPath)) {
+        const tmcLangsResult = await vscode.window.withProgress(
+            { location: vscode.ProgressLocation.Notification, title: "TestMyCode" },
+            async (p) => {
+                return downloadFile(
+                    TMC_JAR,
+                    tmcLangsPath,
+                    undefined,
+                    undefined,
+                    (progress: number, increment: number) =>
+                        p.report({
+                            message: `(${progress}%) Downloading required files`,
+                            increment,
+                        }),
+                );
+            },
+        );
+
+        if (tmcLangsResult.err) {
+            return new Err(tmcLangsResult.val);
+        }
+        console.log("tmc-langs.jar downloaded");
+    }
+
+    const resources: Resources = new Resources(
+        cssPath,
+        extensionVersion,
+        htmlPath,
+        tmcDataPath,
+        tmcLangsPath,
+        tmcWorkspacePath,
+        tmcWorkspaceFilePath,
+        tmcExercisesFolderPath,
+        tmcClosedExercisesFolderPath,
+        mediaPath,
+    );
+
+    return new Ok(resources);
+}
